requirements.txt
streamlit
face_recognition
numpy
opencv-python
Pillow
dlib